Version 1.1
Added a test function in an attempt to find the maximum value of n that can be passed to F() before raising an OverflowError. On a laptop with 6 GB of RAM and 8 GB of swap memory, the test function was able to go up to 594335. This was run in a terminal on Linux.